+++ tests/integration/_cases/mobile_app/mobile_app-upload-help-not-macos.trycmd
@@ -6,8 +6,8 @@ $ sentry-cli mobile-app upload --help
 Usage: sentry-cli[EXE] mobile-app upload [OPTIONS] <PATH>...
 
 Arguments:
-  <PATH>...  The path to the mobile app files to upload. Supported files include Apk, Aab,
-             XCArchive, and IPA.
+  <PATH>...  The path to the mobile app files to upload. Supported files include Apk, Aab, and
+             XCArchive.
 
 Options:
   -o, --org <ORG>
